refactor(usage): extract reporting logic into aghub-usage crate
Move the ccusage shell-out, normalization, and vendor limit-endpoint
logic out of crates/api into a standalone aghub-usage crate so it can be
unit-tested without Rocket or the api binary. The api routes become thin
handlers over aghub_usage::summary / aghub_usage::limits, and the usage
DTOs now live in the new crate (re-exported for ts-rs generation).

- run_ccusage returns String instead of ApiError (no api coupling)
- resolve_ccusage_bin centralizes sidecar/env/PATH lookup for callers
- add unit tests for codex_window parsing and claude/codex normalization
