+++ crates/api/src/dto/mod.rs
@@ -9,4 +9,3 @@ pub mod plugin;
 pub mod skill;
 pub mod sub_agent;
 pub mod transfer;
-pub mod usage;
