+++ crates/usage/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "aghub-usage"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+description = "Usage and rate-limit reporting for AI coding agents"
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
+ts-rs = { workspace = true }
+dirs = { workspace = true }
+tokio = { version = "1", features = [ "process", "time" ] }
+reqwest = { version = "0.13", features = [ "json" ] }
+chrono = "0.4"
+keyring = { version = "3", features = [
+	"apple-native",
+	"windows-native",
+	"linux-native",
+] }
+
+[dev-dependencies]
+tokio = { version = "1", features = [ "process", "time", "rt", "macros" ] }
